refactor: changing code structure to attend mvc pattern
