+++ controller/index.js
@@ -1,0 +1,5 @@
+const userController = require('./userController')
+
+module.exports = {
+  userController
+}
