+++ routes/index.js
@@ -1,0 +1,17 @@
+const { handleError } = require('./../middlewares')
+const userRoute = require('./userRoute')
+
+const router = (app) => {
+  // user routes
+  userRoute(app)
+
+  // error handler
+  app.use((err, _req, res, _next) => {
+    handleError(err, res);
+  })
+}
+
+
+module.exports = {
+  router
+}
